+++ packages/shared-styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @harmony/styles
 
+## 1.0.3
+
+### Patch Changes
+
+- update tailwind config and export
+
 ## 1.0.2
 
 ### Patch Changes