+++ packages/shared-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # shared-ui
 
+## 0.1.2
+
+### Patch Changes
+
+- update shared-ui files
+
 ## 0.1.1
 
 ### Patch Changes